Rename definitions to types
