+++ src/test/language/characterStream.test.ts
@@ -6,8 +6,8 @@ import * as assert from 'assert';
 // tslint:disable-next-line:import-name
 import Char from 'typescript-char';
 import { CharacterStream } from '../../client/language/characterStream';
-import { ICharacterStream, TextRange } from '../../client/language/definitions';
 import { TextIterator } from '../../client/language/textIterator';
+import { ICharacterStream, TextRange } from '../../client/language/types';
 
 // tslint:disable-next-line:max-func-body-length
 suite('Language.CharacterStream', () => {
