+++ src/test/language/textRange.test.ts
@@ -3,7 +3,7 @@
 'use strict';
 
 import * as assert from 'assert';
-import { TextRange } from '../../client/language/definitions';
+import { TextRange } from '../../client/language/types';
 
 // tslint:disable-next-line:max-func-body-length
 suite('Language.TextRange', () => {
